+++ packages/orq-rc/src/models/operations/retrievecontact.ts
@@ -127,7 +127,7 @@ export const RetrieveContactResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-07-15T08:04:17.124Z",
+    "2025-07-16T07:04:13.740Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -165,7 +165,7 @@ export const RetrieveContactResponseBody$outboundSchema: z.ZodType<
   tags: z.array(z.string()).optional(),
   metadata: z.record(z.any()).optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-07-15T08:04:17.124Z"))
+  updated: z.date().default(() => new Date("2025-07-16T07:04:13.740Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {